+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+- Adds [`Objectware.reject`][reject].
+
+[reject]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.reject
+
 ## 1.3.0 (Feb 8, 2015)
 - Adds [`Objectware.isIn`][isIn].